proyectojjsweet/proyectojjsweet/requirements.txt
Flask
psycopg2-binary
gunicorn
reportlab
bcrypt
openpyxl
